refactor: extract restream.rs
